map/python/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.33
pika==1.3.0
